requirements.txt
crewai==0.177.0
crewai-tools>=0.14.0
firecrawl-py==1.6.7
python-dotenv==1.0.1
streamlit==1.37.1
